tiny refactor v2

+++ BlurryRender/headers/mesh.h
@@ -53,7 +53,7 @@ public:
     unsigned int specularNr = 1;
     unsigned int normalNr = 1;
     unsigned int heightNr = 1;
-    for (unsigned int i = 0; i < textures.size(); i++)
+    for (size_t i = 0; i < textures.size(); i++)
     {
       glActiveTexture(GL_TEXTURE0 + i);// active proper texture unit before binding
       // retrieve texture number (the N in diffuse_textureN)
